testcollections, cursor-previous variants, tests may cheat
because of loop for while m.
It makes the test silently leave as soon as m is false
despite what the value for p and v is. If the m is moved
to the always clause everything is tested.

darcs-hash:20070511154905-5228f-3e1644ac0d78d3682dd35f06b6dca8c380f68f49.gz