class-indices-dropped-bugfix
Ignore-this: e620ee90b372b197464cbcee4c96718a

This patch fixes a bug where the get-instances-by-class would return
nil when there were clearly instances in the database (which could be
retrieved using other indices.  We fix this by populating the index
when the controller is opened.  Slow, but it fixes the bug.

Notably, this bug could not be widely reproduced

 ***END OF DESCRIPTION***

Place the long patch description above the ***END OF DESCRIPTION*** marker.
The first line of this file will be the patch name.

This patch contains the following changes:

M ./src/elephant/controller.lisp -2 +4

darcs-hash:20100710193439-513e1-8589483379bd5486f24206ed1a5397dc6c54ee52.gz